# memory: Optimize gum_memory_read for Linux/Android