--- a/gum/backend-linux/gumprocess-linux.c
+++ b/gum/backend-linux/gumprocess-linux.c
@@ -42,6 +42,7 @@
 #include <sys/socket.h>
 #include <sys/syscall.h>
 #include <sys/types.h>
+#include <sys/utsname.h>
 #include <sys/uio.h>
 #include <sys/wait.h>
 #ifdef HAVE_SYS_USER_H
@@ -3582,3 +3583,37 @@ gum_libc_syscall_4 (gsize n,
 
   return result;
 }
+
+gboolean
+gum_linux_check_kernel_version (guint major,
+                               guint minor,
+                               guint micro)
+{
+  static gboolean initialized = FALSE;
+  static guint kernel_major = G_MAXUINT;
+  static guint kernel_minor = G_MAXUINT;
+  static guint kernel_micro = G_MAXUINT;
+
+  if (!initialized)
+  {
+    struct utsname un;
+    G_GNUC_UNUSED int res;
+
+    res = uname (&un);
+    g_assert (res == 0);
+
+    /* Linux version string format: "X.Y.Z-extra" */
+    sscanf (un.release, "%u.%u.%u", &kernel_major, &kernel_minor, &kernel_micro);
+
+    initialized = TRUE;
+  }
+
+  if (kernel_major > major)
+    return TRUE;
+
+  if (kernel_major == major && kernel_minor > minor)
+    return TRUE;
+
+  return kernel_major == major && kernel_minor == minor && kernel_micro >= micro;
+}
+